Resolves: #89 Stored Credential Support
This adds support for Stored Credentials on Authorization and Purchase requests.
