+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- @jeremywrowe - Add Support for Stored Credentials
 - @jeremywrowe - Add 3D Secure 2 abilities to gateway class
 
 ## [2.0.20] - 2019-06-27
