+++ lib/spreedly.rb
@@ -13,6 +13,8 @@ require 'spreedly/payment_methods/paypal'
 require 'spreedly/payment_methods/sprel'
 require 'spreedly/payment_methods/bank_account'
 require 'spreedly/payment_methods/third_party_token'
+require 'spreedly/transactions/shipping_address'
+require 'spreedly/transactions/response'
 require 'spreedly/transactions/transaction'
 require 'spreedly/transactions/gateway_transaction'
 require 'spreedly/transactions/add_payment_method'
